API/Express: separating app entry codes
