+++ API/CarDB-Express/src/index.js
@@ -1,28 +1,22 @@
-require('dotenv').config()
-const express = require('express');
-const cors = require('cors');
-const routes = require("./CarsRouter.js");
+const { swaggerDocs } = require("./swagger");
+const app = require("./app");
 
 const PORT = process.env.PORT || 3000;
-const app = express();
 
-if(process.env.USE_API === "firestore"){
-    const {connect, initData} = require("./firestore/db-init.js");
-    connect(); // Initialize Firebase
-    initData(); // Initialize data
-}else{
-    const {initDB} = require("./sqlite/db.js")
-    initDB();
+let server = app.listen(PORT, () => {
+    console.log('HTTP server is running on Port', PORT)
+    swaggerDocs(app, PORT);
+})
+
+
+const close = () => {
+  console.log('Gracefully stopping...')
+  server.close(async () => {
+    console.log(`HTTP${env.HTTPS ? 'S' : ''} server closed`)
+    await databaseHelper.close(true)
+    console.log('Database connection closed')
+    process.exit(0)
+  })
 }
 
-app.use(cors());
-app.use(express.json());
-
-app.use("/api/cars/", routes);
-
-const { swaggerDocs } = require("./swagger");
-
-app.listen(PORT, () => {
-    console.log(`Server is running on http://localhost:${PORT}`);
-    swaggerDocs(app, PORT);
-});
+['SIGINT', 'SIGTERM', 'SIGQUIT'].forEach((signal) => process.on(signal, close))
